+++ apps/builder/app/builder/features/props-panel/props-panel.tsx
@@ -19,7 +19,7 @@ import {
 import { ChevronDownIcon } from "@webstudio-is/icons";
 import type { Publish } from "~/shared/pubsub";
 import { propsStore, useInstanceProps } from "~/shared/nano-states";
-import { CollapsibleSectionWithAddButton } from "~/builder/shared/inspector";
+import { CollapsibleSectionWithAddButton } from "~/builder/shared/collapsible-section";
 import {
   useStyleData,
   type SetProperty as SetCssProperty,
