+++ apps/builder/app/builder/shared/collapsible-section.tsx
