--- apps/builder/app/builder/shared/inspector/index.tsx
@@ -1,2 +1,0 @@
-export * from "./collapsible-section";
-export * from "./component-info";
